feat: add Facebook login functionality
Add support for Facebook login by installing the `react-facebook` package and
adding the `REACT_APP_FACEBOOK_APP_ID` environment variable to the `.env` file.
Create a new `Login` component that uses the `useLogin` hook from `react-facebook`
to handle the login process. Add a new route to the `routes.js` file for the
`Login` component. Modify the `SliderMain` component to add keys to the buttons
in the `items` array.